requirements.txt
Django==1.10.1
PyYAML==3.11
django-appconf==1.0.2
django-cors-headers
django-datetime-widget
django-etc
django-markdown-deux==1.0.5
django-tagulous
djangorestframework-jwt
djangorestframework==3.5.0
etc
git+ssh://git@github.tamu.edu/cpt/hashids-py.git
pinax
pinax-announcements==2.0.3
psycopg2
pytz==2016.4
six
django-filter
